utils: merge date and datefilter
